Back/requirements.txt
blinker==1.7.0
click==8.1.7
colorama==0.4.6
distlib==0.3.8
filelock==3.13.1
Flask==3.0.2
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
pip==24.0
platformdirs==4.2.0
PyMySQL==1.1.0
setuptools==69.2.0
virtualenv==20.25.1
Werkzeug==3.0.1
PyJWT==2.8.0
pytz==2024.1
#  Installation: pip install -r ./requirements.txt